Miriteam/prod/prod/requirements.txt
opencv_python==4.2.0.34
SpeechRecognition==3.8.1
Pillow==7.1.2
